refactor: netsim
