+++ .gitignore
@@ -8,3 +8,4 @@ netsim/data
 netsim/keys
 netsim/bins
 netsim/viz
+.vscode
