+++ Cargo.toml
@@ -1,37 +1,4 @@
-[package]
-name = "chuck"
-version = "0.0.1"
-edition = "2021"
-authors = ["arqu <asmir@n0.computer>"]
-license = "Apache-2.0/MIT"
-repository = "https://github.com/n0-computer/chuck"
-description = "iroh test tools"
-rust-version = "1.65"
+[workspace]
+members = ["chuck"]
 
-[dependencies]
-anyhow = { version = "1", features = ["backtrace"] }
-clap = { version = "4.0.9", features = ["derive"] }
-tokio = { version = "1", features = ["full"] }
-serde = { version = "1", features = ["derive"] }
-bincode = "1.3.3"
-tempfile = "3.4.0"
-futures = "0.3.21"
-bytes = "1.1.0"
-rand = "0.8.5"
-
-axum = "0.6.2"
-tower = { version = "0.4", features = ["util"] }
-tower-http = { version = "0.3.0", features = ["fs", "trace"] }
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-axum-server = { version = "0.3", features = ["tls-rustls"] }
-reqwest = { version = "0.11.10", default-features = false, features = ["rustls-tls"] }
-
-#iroh-api = { git = "https://github.com/n0-computer/iroh", rev = "0e06dd7f70e21955f735ecd54464c7527fbf72c1"}
-#iroh-util = { git = "https://github.com/n0-computer/iroh", rev = "0e06dd7f70e21955f735ecd54464c7527fbf72c1"}
-#iroh-share = { git = "https://github.com/n0-computer/iroh", rev = "0e06dd7f70e21955f735ecd54464c7527fbf72c1"}
-#iroh-one = { git = "https://github.com/n0-computer/iroh", rev = "0e06dd7f70e21955f735ecd54464c7527fbf72c1"}
-#iroh-rpc-types = { git = "https://github.com/n0-computer/iroh", rev = "0e06dd7f70e21955f735ecd54464c7527fbf72c1", default-features = false}
-#iroh-rpc-client = { git = "https://github.com/n0-computer/iroh", rev = "0e06dd7f70e21955f735ecd54464c7527fbf72c1", default-features = false}
-
-iroh = { git = "https://github.com/n0-computer/iroh", rev = "2149bc8c6819b044833f5af0b8fdac567ef93650", default-features = false }
+resolver = "2"
