+++ netsim/scripts/python_deps.sh
@@ -1,0 +1,1 @@
+pip3 install pyshark drawsvg dpkt humanfriendly
