+++ netsim/scripts/ubuntu_deps.sh
@@ -1,0 +1,6 @@
+sudo apt install mininet
+sudo apt install openvswitch-testcontroller
+sudo apt install iperf
+sudo apt install tshark
+sudo apt install smcroute
+sudo systemctl enable smcroute.service
