+++ netsim/sims/iroh/iroh.json
@@ -6,7 +6,7 @@
             "description": "",
             "nodes": [
                 {
-                    "name": "iroh_srv",
+                    "name": "i_srv",
                     "count": 1,
                     "cmd": "./bins/iroh start --add data/1G.bin",
                     "type": "public",
@@ -17,13 +17,13 @@
                     "param_parser": "iroh_ticket"
                 },
                 {
-                    "name": "iroh_get",
+                    "name": "i_get",
                     "count": 1,
                     "cmd": "time ./bins/iroh blobs get --start %s --out STDOUT > /dev/null",
                     "type": "public",
                     "connect": {
                         "strategy": "params",
-                        "node": "iroh_srv"
+                        "node": "i_srv"
                     },
                     "process": "short",
                     "parser": "iroh_1gb"
@@ -35,7 +35,7 @@
             "description": "",
             "nodes": [
                 {
-                    "name": "iroh_srv",
+                    "name": "i_srv",
                     "count": 1,
                     "cmd": "./bins/iroh start --add data/1G.bin",
                     "type": "public",
@@ -46,13 +46,13 @@
                     "param_parser": "iroh_ticket"
                 },
                 {
-                    "name": "iroh_get",
+                    "name": "i_get",
                     "count": 3,
                     "cmd": "time ./bins/iroh blobs get --start %s --out STDOUT > /dev/null",
                     "type": "public",
                     "connect": {
                         "strategy": "params",
-                        "node": "iroh_srv"
+                        "node": "i_srv"
                     },
                     "process": "short",
                     "parser": "iroh_1gb"
@@ -64,7 +64,7 @@
             "description": "",
             "nodes": [
                 {
-                    "name": "iroh_srv",
+                    "name": "i_srv",
                     "count": 1,
                     "cmd": "./bins/iroh start --add data/1G.bin",
                     "type": "public",
@@ -75,13 +75,13 @@
                     "param_parser": "iroh_ticket"
                 },
                 {
-                    "name": "iroh_get",
+                    "name": "i_get",
                     "count": 5,
                     "cmd": "time ./bins/iroh blobs get --start %s --out STDOUT > /dev/null",
                     "type": "public",
                     "connect": {
                         "strategy": "params",
-                        "node": "iroh_srv"
+                        "node": "i_srv"
                     },
                     "process": "short",
                     "parser": "iroh_1gb"
@@ -93,7 +93,7 @@
             "description": "",
             "nodes": [
                 {
-                    "name": "iroh_srv",
+                    "name": "i_srv",
                     "count": 1,
                     "cmd": "./bins/iroh start --add data/1G.bin",
                     "type": "public",
@@ -104,13 +104,13 @@
                     "param_parser": "iroh_ticket"
                 },
                 {
-                    "name": "iroh_get",
+                    "name": "i_get",
                     "count": 10,
                     "cmd": "time ./bins/iroh blobs get --start %s --out STDOUT > /dev/null",
                     "type": "public",
                     "connect": {
                         "strategy": "params",
-                        "node": "iroh_srv"
+                        "node": "i_srv"
                     },
                     "process": "short",
                     "parser": "iroh_1gb"
@@ -122,7 +122,7 @@
             "description": "",
             "nodes": [
                 {
-                    "name": "iroh_srv",
+                    "name": "i_srv",
                     "count": 2,
                     "cmd": "./bins/iroh start --add data/1G.bin",
                     "type": "public",
@@ -133,13 +133,13 @@
                     "param_parser": "iroh_ticket"
                 },
                 {
-                    "name": "iroh_get",
+                    "name": "i_get",
                     "count": 2,
                     "cmd": "time ./bins/iroh blobs get --start %s --out STDOUT > /dev/null",
                     "type": "public",
                     "connect": {
                         "strategy": "params",
-                        "node": "iroh_srv"
+                        "node": "i_srv"
                     },
                     "process": "short",
                     "parser": "iroh_1gb"
@@ -151,7 +151,7 @@
             "description": "",
             "nodes": [
                 {
-                    "name": "iroh_srv",
+                    "name": "i_srv",
                     "count": 2,
                     "cmd": "./bins/iroh start --add data/1G.bin",
                     "type": "public",
@@ -162,13 +162,13 @@
                     "param_parser": "iroh_ticket"
                 },
                 {
-                    "name": "iroh_get",
+                    "name": "i_get",
                     "count": 4,
                     "cmd": "time ./bins/iroh blobs get --start %s --out STDOUT > /dev/null",
                     "type": "public",
                     "connect": {
                         "strategy": "params",
-                        "node": "iroh_srv"
+                        "node": "i_srv"
                     },
                     "process": "short",
                     "parser": "iroh_1gb"
@@ -180,7 +180,7 @@
             "description": "",
             "nodes": [
                 {
-                    "name": "iroh_srv",
+                    "name": "i_srv",
                     "count": 2,
                     "cmd": "./bins/iroh start --add data/1G.bin",
                     "type": "public",
@@ -191,13 +191,13 @@
                     "param_parser": "iroh_ticket"
                 },
                 {
-                    "name": "iroh_get",
+                    "name": "i_get",
                     "count": 6,
                     "cmd": "time ./bins/iroh blobs get --start %s --out STDOUT > /dev/null",
                     "type": "public",
                     "connect": {
                         "strategy": "params",
-                        "node": "iroh_srv"
+                        "node": "i_srv"
                     },
                     "process": "short",
                     "parser": "iroh_1gb"
@@ -209,7 +209,7 @@
             "description": "",
             "nodes": [
                 {
-                    "name": "iroh_srv",
+                    "name": "i_srv",
                     "count": 2,
                     "cmd": "./bins/iroh start --add data/1G.bin",
                     "type": "public",
@@ -220,13 +220,13 @@
                     "param_parser": "iroh_ticket"
                 },
                 {
-                    "name": "iroh_get",
+                    "name": "i_get",
                     "count": 10,
                     "cmd": "time ./bins/iroh blobs get --start %s --out STDOUT > /dev/null",
                     "type": "public",
                     "connect": {
                         "strategy": "params",
-                        "node": "iroh_srv"
+                        "node": "i_srv"
                     },
                     "process": "short",
                     "parser": "iroh_1gb"
